fix: broken media function

+++ include/squirrel_constructors.h
@@ -6,6 +6,6 @@
 
 struct key make_key(uint8_t keycode);
 struct key make_modifier(uint8_t modifier_code);
-struct key make_media(uint8_t media_code);
+struct key make_media(uint16_t media_code);
 
 #endif
